+++ Microcontroller/Magnetometer.h
@@ -58,5 +58,5 @@
 class Magnetometer {
     public:
         void start();
-        void get_values(int16_t* readings);
+        void get_values(uint8_t* readings);
 };
